add helpers for fonts etc
